+++ tests/deprecated_api/test_remove_1-8.py
@@ -140,6 +140,24 @@ def test_v1_8_0_deprecated_trainer_should_rank_save_checkpoint(tmpdir):
         _ = trainer.should_rank_save_checkpoint
 
 
+def test_v1_8_0_trainer_optimizers_mixin():
+    trainer = Trainer()
+    model = BoringModel()
+    trainer.strategy.connect(model)
+    trainer.lightning_module.trainer = trainer
+
+    with pytest.deprecated_call(
+        match=r"`TrainerOptimizersMixin.init_optimizers` was deprecated in v1.6 and will be removed in v1.8."
+    ):
+        trainer.init_optimizers(model)
+
+    with pytest.deprecated_call(
+        match=r"`TrainerOptimizersMixin.convert_to_lightning_optimizers` was deprecated in v1.6 and will be removed in "
+        "v1.8."
+    ):
+        trainer.convert_to_lightning_optimizers()
+
+
 def test_v1_8_0_deprecate_trainer_callback_hook_mixin():
     methods_with_self = [
         "on_before_accelerator_backend_setup",